requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.0
Deprecated==1.2.15
executing==2.1.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
logfire==2.7.1
markdown-it-py==3.0.0
mdurl==0.1.2
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-requests==0.50b0
opentelemetry-instrumentation-sqlalchemy==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
packaging==24.2
protobuf==5.29.1
psycopg2-binary==2.9.10
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
SQLAlchemy==2.0.36
tinydb==4.8.2
typing_extensions==4.12.2
urllib3==2.2.3
wrapt==1.17.0
zipp==3.21.0
